requirements.txt
langchain>=0.1.0
langgraph>=0.0.15
python-dotenv>=1.0.0
groq>=0.3.0
pydot>=1.4.2
graphviz>=0.20.1
mermaid-cli>=0.1.12  # For DAG diagram generation
